+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The releases of major and minor versions contain an overview of changes since th
 Version 1.6.x
 -------------
 
+## Version 1.6.3 (20xx/xx)
+- Added support for multi-objective model checking of long-run average objectives including mixtures with other kinds of objectives.
+- Added support for generating optimal schedulers for globally formulae
+- Simulator supports exact arithmetic
+- `storm-pomdp`: States can be labelled with values for observable predicates
+- `storm-pomdp`: (Only API) Track state estimates
+- `storm-pomdp`: (Only API) Reduce computation of state estimates to computation on unrolled MDP
+
 ## Version 1.6.2 (2020/09)
 - Prism program simplification improved.
 - Revamped implementation of long-run-average algorithms, including scheduler export for LRA properties on Markov automata.